refactor lattices
